refactor: reorganize ui toolkit files
