+++ Assets/Scripts/Abstract Classes/EntityClass.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using UI_Toolkit;
 using UnityEngine;
 using static CardComparator;
 using static StatusEffect;
