+++ Assets/Scripts/UI Elements/BattleQueueIcons.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
+using UI_Toolkit;
 using UnityEngine;
 
 public class BattleQueueIcons : DisplayableClass
