+++ Assets/Scripts/VNScripts/DialogueManager.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
+using UI_Toolkit;
 using UnityEngine;
 
 #nullable enable
